refactoring to cleanup and better understand what needs to be done
